cleaner with cost benefit ratio

+++ headers/cost_benefit.hpp
@@ -3,5 +3,5 @@
 class CostBenefit {
     public:
         unsigned int segmentNumber;
-        unsigned int score;
+        double score;
 };
